+++ src/pygnssutils/gnssserver.py
@@ -23,7 +23,13 @@ from threading import Thread
 from time import sleep
 from typing import Literal
 
-from pygnssutils.globals import NTRIP2, OUTPORT, PYGNSSUTILS_PEM, PYGNSSUTILS_PEMPATH
+from pygnssutils.globals import (
+    NTRIP2,
+    OUTPORT,
+    PYGNSSUTILS_PEM,
+    PYGNSSUTILS_PEMPATH,
+    RTCMSTR,
+)
 from pygnssutils.gnssstreamer import GNSSStreamer
 from pygnssutils.helpers import format_conn, ipprot2int
 from pygnssutils.socket_server import ClientHandler, ClientHandlerTLS, SocketServer
@@ -43,11 +49,12 @@ class GNSSSocketServer:
         outport: int = OUTPORT,
         tls: bool = False,
         maxclients: int = 5,
-        ntripmode: int = 0,
+        ntripmode: Literal[0, 1] = 0,
         ntripversion: Literal["1.0", "2.0"] = NTRIP2,
         ntripuser: str = "anon",
         ntrippassword: str = "password",
         tlspempath: str = getenv(PYGNSSUTILS_PEMPATH, PYGNSSUTILS_PEM),
+        ntriprtcmstr: str = RTCMSTR,
         **kwargs,
     ):
         """
@@ -64,11 +71,12 @@ class GNSSSocketServer:
         :param int outport: TCP port (50010)
         :param bool tls: Enable TLS (HTTPS) (False)
         :param int maxclients: maximum number of connected clients (5)
-        :param int ntripmode: 0 = socket server, 1 - NTRIP server (0)
+        :param Literal[0,1] ntripmode: 0 = socket server, 1 - NTRIP server (0)
         :param Literal["1.0","2.0"] ntripversion: NTRIP version "1.0"/"2.0" ("2.0")
         :param str ntripuser: NTRIP caster authentication user ("anon")
         :param str ntrippassword: NTRIP caster authentication password ("password")
         :param str tlspempath: Path to TLS PEM file ("pygnssutils.pem")
+        :param str ntriprtcmstr: NTRIP caster RTCM types sourcetable entry e.g. '1006(5),1077(1),...'
         :param dict kwargs: optional keyword arguments to pass to GNSSStreamer
         """
 
@@ -85,6 +93,7 @@ class GNSSSocketServer:
             self._ntripuser = ntripuser
             self._ntrippassword = ntrippassword
             self._tlspempath = tlspempath
+            self._ntriprtcmstr = ntriprtcmstr
             self._hostip = hostip
             self._outport = int(outport)
             self._tls = tls
@@ -192,6 +201,7 @@ class GNSSSocketServer:
                 self._ntripuser,
                 self._ntrippassword,
                 self._tlspempath,
+                self._ntriprtcmstr,
             ),
             daemon=True,
         )
@@ -219,13 +229,14 @@ class GNSSSocketServer:
         hostip: str,
         outport: int,
         tls: bool,
-        ntripmode: int,
+        ntripmode: Literal[0, 1],
         maxclients: int,
         output: object,
-        ntripversion: int,
+        ntripversion: Literal["1.0", "2.0"],
         ntripuser: str,
         ntrippassword: str,
         tlspempath: str,
+        ntriprtcmstr: str,
     ):
         """
         THREADED
@@ -237,8 +248,8 @@ class GNSSSocketServer:
         :param str outport: TCP port
         :param bool tls: Use TLS
         :param int maxclients: maximum number of connected clients
-        :param int ntripmode:
-        :param str ntripversion: NTRIP version
+        :param Literal[0,1] ntripmode:
+        :param Literal["1.0","2.0"] ntripversion: NTRIP version
         :param str ntripuser: NTRIP caster authentication user
         :param str ntrippassword: NTRIP caster authentication password
         :param str tlspempath: path to TLS PEM file
@@ -259,6 +270,7 @@ class GNSSSocketServer:
                 ntrippassword=ntrippassword,
                 ipprot=ipprot,
                 tlspempath=tlspempath,
+                ntriprtcmstr=ntriprtcmstr,
             ) as self._socket_server:
                 self._socket_server.serve_forever()
         except OSError as err: